# MAINT,ENH: Rewrite scalar math logic