--- a/numpy/core/src/multiarray/abstractdtypes.c
+++ b/numpy/core/src/multiarray/abstractdtypes.c
@@ -259,6 +259,7 @@ NPY_NO_EXPORT PyArray_DTypeMeta PyArray_PyIntAbstractDType = {{{
         .tp_name = "numpy._IntegerAbstractDType",
     },},
     .flags = NPY_DT_ABSTRACT,
+    .type_num = -1,
     .dt_slots = &pyintabstractdtype_slots,
 };
 
@@ -276,6 +277,7 @@ NPY_NO_EXPORT PyArray_DTypeMeta PyArray_PyFloatAbstractDType = {{{
         .tp_name = "numpy._FloatAbstractDType",
     },},
     .flags = NPY_DT_ABSTRACT,
+    .type_num = -1,
     .dt_slots = &pyfloatabstractdtype_slots,
 };
 
@@ -293,5 +295,6 @@ NPY_NO_EXPORT PyArray_DTypeMeta PyArray_PyComplexAbstractDType = {{{
         .tp_name = "numpy._ComplexAbstractDType",
     },},
     .flags = NPY_DT_ABSTRACT,
+    .type_num = -1,
     .dt_slots = &pycomplexabstractdtype_slots,
 };