… benchmarks/benchmarks/bench_scalar.py
… doc/release/upcoming_changes/21188.performance.rst
… numpy/core/setup_common.py
… numpy/core/tests/test_scalarmath.py