This tries to redo the scalar math logic to take some more care about subclasses, but most of all introduce logic to defer to the `other` if `self` can be cast to it safely (i.e. it is the correct promotion).

This makes things much faster and more reliable, since we now use defer much less to `ufuncs` indirectly.  This ensures that integer overflows are reported more reliably.

Another major point about it, is that this reorganizes the coercion of Python int, float, complex (and bool).  This should help a bit with switching to "weak" Python scalars.
    
Further, it contains a commit to make all macros to inline functions and move the floating point overflow flag handling to a return value.  Checking floating point flags is not insanely slow, but it is pretty slow on the scale of integer operations here (~30% on my computer).

It should fix some bugs around subclasses, but then subclassing scalars should still be pretty prone to issues (similar to subclassing arrays I guess).  Complicated subclass cases may still end up in the generic (array) path, but we catch a few more ahead of time.

---

This is the 2-3 approach I have been thinking about.  Another one would be to have (more or less) a single function dealing with any scalar inputs.  It does "inline" the casting logic here.  I do not like that, but it seemed somewhat straight forward – it would be nice to create `npy_cast_@from@_to_@to@` or so functions, to use here more generically.

The alternative would be logic with the existing "cast" functionality, but it should be slower, and while verbose (and the current macros are ugly), I am not sure that would actually end up much better.

The `PyArray_ScalarFromObject` seemed only useful for the old scalar paths, so I added a deprecation:  Neither were even documented, and both would probably need some work to transition to the new DTypes well.
`PyArray_CastScalarDirect` may also be a target, but it is still used in at least one place, so likely should just fix it.

---

Benchmarks:
```
       before           after         ratio
     [732ed25e]       [3b60effc]
     <main>           <scalar-math-rewrite>
-        861±10ns          819±6ns     0.95  bench_scalar.ScalarMath.time_abs('int32')
-         862±3ns          819±5ns     0.95  bench_scalar.ScalarMath.time_abs('float16')
-        929±10ns         882±10ns     0.95  bench_scalar.ScalarMath.time_addition('complex128')
-     1.05±0.02μs          996±4ns     0.95  bench_scalar.ScalarMath.time_addition('float16')
-         859±4ns          813±3ns     0.95  bench_scalar.ScalarMath.time_abs('int16')
-         890±8ns          841±2ns     0.94  bench_scalar.ScalarMath.time_abs('complex64')
-         924±4ns          871±4ns     0.94  bench_scalar.ScalarMath.time_multiplication('float64')
-         957±2ns        895±0.8ns     0.93  bench_scalar.ScalarMath.time_abs('complex256')
-        932±10ns        871±0.4ns     0.93  bench_scalar.ScalarMath.time_addition('float64')
-         922±5ns        855±0.8ns     0.93  bench_scalar.ScalarMath.time_abs('longfloat')
-        893±10ns          826±7ns     0.93  bench_scalar.ScalarMath.time_abs('complex128')
-     1.09±0.07μs          987±8ns     0.90  bench_scalar.ScalarMath.time_multiplication('float16')
-        687±10ns          532±2ns     0.77  bench_scalar.ScalarMath.time_add_int32_other('int32')
-     4.63±0.05μs       3.58±0.1μs     0.77  bench_scalar.ScalarMath.time_addition_pyint('int16')
-     5.58±0.06μs      4.24±0.07μs     0.76  bench_scalar.ScalarMath.time_addition_pyint('float32')
-     5.62±0.08μs       4.22±0.1μs     0.75  bench_scalar.ScalarMath.time_addition_pyint('complex64')
-     1.03±0.01μs          755±9ns     0.73  bench_scalar.ScalarMath.time_add_int32_other('int16')
-     5.76±0.03μs       4.21±0.1μs     0.73  bench_scalar.ScalarMath.time_addition_pyint('float16')
-         922±2ns          662±2ns     0.72  bench_scalar.ScalarMath.time_addition('int16')
-         921±3ns        653±0.5ns     0.71  bench_scalar.ScalarMath.time_multiplication('int64')
-         929±1ns          656±6ns     0.71  bench_scalar.ScalarMath.time_addition('int64')
-         916±8ns          645±4ns     0.70  bench_scalar.ScalarMath.time_multiplication('int16')
-        953±60ns          657±5ns     0.69  bench_scalar.ScalarMath.time_multiplication('int32')
-         934±8ns        642±0.9ns     0.69  bench_scalar.ScalarMath.time_addition('int32')
-     3.54±0.01μs         1.76±0μs     0.50  bench_scalar.ScalarMath.time_power_of_two('longfloat')
-     3.41±0.04μs      1.59±0.01μs     0.47  bench_scalar.ScalarMath.time_power_of_two('float64')
-      3.70±0.1μs         1.65±0μs     0.45  bench_scalar.ScalarMath.time_power_of_two('complex256')
-     2.36±0.05μs         1.04±0μs     0.44  bench_scalar.ScalarMath.time_power_of_two('int64')
-     2.83±0.05μs         1.19±0μs     0.42  bench_scalar.ScalarMath.time_addition_pyint('longfloat')
-     2.87±0.05μs         1.20±0μs     0.42  bench_scalar.ScalarMath.time_addition_pyint('complex256')
-     3.17±0.08μs         1.25±0μs     0.39  bench_scalar.ScalarMath.time_power_of_two('complex128')
-     2.64±0.02μs          992±5ns     0.38  bench_scalar.ScalarMath.time_addition_pyint('complex128')
-     2.17±0.04μs          807±2ns     0.37  bench_scalar.ScalarMath.time_addition_pyint('int64')
-     2.75±0.02μs          993±5ns     0.36  bench_scalar.ScalarMath.time_addition_pyint('float64')
-      14.1±0.2μs      1.24±0.03μs     0.09  bench_scalar.ScalarMath.time_add_int32_other('longfloat')
-      14.6±0.3μs      1.23±0.01μs     0.08  bench_scalar.ScalarMath.time_add_int32_other('complex256')
-      14.1±0.2μs      1.10±0.01μs     0.08  bench_scalar.ScalarMath.time_add_int32_other('float64')
-      14.5±0.1μs      1.11±0.01μs     0.08  bench_scalar.ScalarMath.time_add_int32_other('complex128')
-     13.9±0.07μs          973±6ns     0.07  bench_scalar.ScalarMath.time_add_int32_other('int64')
```
(Some corner cases may be significantly slower, mainly certain `scalar + 0d_array` ops, but I am not sure I want to worry about those much.)

---

**This PR became quite big, I may split it up.  At this time it relies on gh-21178.   It is currently missing some additional tests for the subclass behavior at least, but I would like to check code-coverage on that front as well.**