@seberg 2022-03-14
Hmmmpf, somehow going to `static inline` functions makes clang optimize some floating point errors away for some multiplications.