@seberg 2022-03-18
Puh... Added (tricky) tests and fixed some other bugs:
* float16 division was just plain crashing
* small integer division goes to float64 in the ufunc machinery, but went to float32 here.  (It may well have been what division did at some point)

I did not fix that complex comparisons behave differently for arrays and scalars (ufuncs use NaN aware order currently?!).  It is still a change, since the scalar path will be used  a bit more often, but...

Otherwise, these are the scalar-math changes as discussed @mattip.  It is still based off gh-21178, though.