@seberg 2022-04-24
There are currently two issues.  The first is that on clang+mac, some floating point errors are not reported.  I really don't quite understand why.  Locally on linux (clang 11 though) it doesn't happen (also with `-O3`).  I have tried to use `FINLINE`, since the main change should be the change from macro to function.
(Another try now is adding `volatile` to all of the math `out` functions, but basically, I am out of ideas besides looking at the generated instructions to see what the compiler is doing.  The simple try with godbolt didn't reall help there either.)

The other is that PyPy seems to behave slightly different for `tp_richcompare`, I suspect that is just OK, but should check it.