@seberg 2022-04-28
Btw. when it comes to the floating point error issues on clang... It turns out the problem was moving the the `fpe_clear_floatstatus_barrier` calls lower.  Somehow using `out` "barrier" seems to have not been enough to ensure the correct order.

Considering that it seemed like it _sometimes_ was correct, I am even wondering if FPEs can be checked out-of-band and it is possible to create race conditions...  But that is more of a curiosity, I am not worried that the fix of moving the FPE checking is not reasonable.