@seberg 2022-05-04
Ping @mattip, not sure how to proceed.  Do you want to have another look pair programming or can we go ahead?
Or maybe @ganesh-k13 can also have a look, since he had once started looking at similar things.