@mattip 2022-05-06
LGTM. The refactoring looks OK, and the speedups are great.

I assume the benchmarks at the top of the PR are still valid?
> Some corner cases may be significantly slower, mainly certain scalar + 0d_array ops ...

Could you add some benchmarks for these and quantify the slowdown?