@seberg 2022-05-06
The big slowdown is that for:
```
arr = np.array(1.)
scalar = np.float64(1.)
```
when the array can be cast to the scalar:
```
%timeit scalar + arr
# 287 ns ± 0.303 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```
Used to be much faster (scalar path kicked in), then:
```
%timeit arr + scalar
# 1.67 µs ± 9.22 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```
And for now they both go through the array path.  So this is a huge slowdown in that very specific case.  The `scalar + arr` case there is still 1/3 slower than `scalar + scalar`, since we used to convert the 0-D array to a scalar, and only then try to continue.

I have added a benchmark for it.