@mattip 2022-05-06
> So this is a huge slowdown in that very specific case

Sorry, I am not following. What was the new benchmark performance before the PR and what is it afterwards? How does the time for the new benchmarks `time_add_int32arr_and_other` compare to `time_add_int32_other` ?