@seberg 2022-05-06
> We may need to back it out if it turns out 0d arrays are used more than we think, but I agree scalars are likely to be much more common.

Yeah, or re-add special cases for 0-D arrays.  Although, in that case it would be nice to at least also make `arr + scalar` fast and not just `scalar + arr`.  The assymetry adds to my feeling of not wanting to worry much about it.